requirements/dev.txt
flake8
bpython
djangorestframework
markdown
django-filter
